Add FAQ entry for GRIT
In the games I've played, we always have questions around remembering what exactly GRIT does and doesn't apply to. I just compiled it all here, because I'm hoping it will be helpful to have this all compiled in a single entry, since the full rules around GRIT are a bit scattered in the core book.